Section 2 - Player Setup
Created a new CharacterBody3D scene for the Player Character. Added a spring arm and camera so the Player is shown in the Game scene when the game starts. Added a script to the Player and edited the script to allow for basic movement. Added an Input Map to the editor and setup rotation of the the Player in the direction of movement. Added an OmniLight to the Player to better see him in the scene. Also implemented Idle and Run animations to play when the Player is standing still and moving, respectively. Finally I added swords to the Player's hands that will be used for attacking later in the game.