Section 4 - Collectable Coin Pickup
Created a Coin pickup with visual effects that play when the Coin is picked up by the Player. Also created an animation that shows the coin flying up in the air and eventually disappears. Also created a UI that shows the number of coins collected by the Player anchored to the upper right hand corner of the screen.