Section 5 - Create the Enemy AI using Navigation
Created an Enemy scene with basic movement using both NavigationRegion3d (to see the path) and NavigationAgent3D (to walk the path). Wrote a basic Enemy script that uses both the "walk" and "idle" animations when appropriate. Also added direction code so that the Enemy rotates towards the Player when walking. Eventually, this basic code will be replaced by Enemy States.