Section 12 - Implementing Enemy States
Used the basis of the State Machine to parent 4 basic states for the Enemy: Spawn, ChasePlayer, Attack and Dead. Spawn State is the initial state and goes into effect when the Enemy first appears in the game. ChasePlayer happens immediate after Spawn when the Enemy follows the Player. Attack happens when the Enemy is close enough to hit the Player. Findally, Dead state is engaged with the health of the Enemy is equal to zero. Animations where also include when Spawning, Attacking the Player and when Dead.