Section 13 - The Player Health
Added a variable for the Player health and added two more states to the Player's State Machine: Hurt & Dead. Added animation effects to the Hurt state so that the Player plays an injured animation and flashes red when hit by the Enemy. Dead state also plays an animation of the Player falling to the ground. Finally, added a Player health bar to the UI scene which gets updated every time the Player enters the Hurt State. Now with both character's finished (Player and Enemy) we will focus on other areas of the game in upcoming commits.