Step 14 - The Healing Orb
Created a Healing Orb Scene in the game and added a script that plays a visual effect when picked up, restores a set variable of health to the Player and removes itself from the game. Last thing we did in this commit was to update the Enemy Dead State to drop the healing orb after it has been killed.