Section 16 - Enemy Spawner & Level Design
Created a Level scene that contains two to three Spawn Points, a Spawn Trigger and a Gate scene. Added a script to the Level scene that spawns an Enemy at each Spawn Point when the Player overlaps with the Spawn Trigger. Updated the Gate script so that it opens after each Enemy in a Level is destroyed (null). Finally, added a Portal Scene with a script that allows the Player to pass if they have enough coins to proceed.