Section 17 - Game Play UI
Added another Control node to the UI scene to manage three different screens, Game Over, Game Finish and Game Pause. Moved the GameUI script from the first Control node to the root UI node and adjusted the script to accommodate the new node structure. Obviously, we'll be adding to this script to control the features of the added Control node.